Some fixes to the website
I've tried to pull in links which pointed to the old GNA website and localize them to the GitHub pages by copying those files into the directory structure here.  Ideally, this should lower the search ranking of the old website by reducing the number of links to it while increasing the search ranking of the new site by increasing the number of links to its pages.  Of course, given the small number of links involved, this may not have an immediately noticeable effect.
I've also tried to clarify the download and installation instructions for Mac and Linux regarding what to download and what the prerequisites are.