Links in header need to be absolute
Relative links in html are relative to the page which is open, not the location of the file the link was derived from.  As a result, the links in the menu need to be absolute because files in different relative locations use the same header.

Note: this means that the links won't work when simply opening the local files with your web browser.  To properly test you'll need to be running a local web server.  This can be done with Python 2.7 by executing `python -m SimpleHTTPServer 8888 &` from the repository directory in Terminal.  Your browser will then find the website at `http://localhost:8888/`