Make image size dependent on resolution and not fixed
By tuning the resolution down to just what is needed to get the size images I want in the tutorial, I can eliminate the resize argument, allowing the images created to be just as big as they need to be.  This allows the script to generate some of the smaller images.