Reworking gabc file tutorial
Instead of using the `\greannotation` and `\grecommentary` commands directly, I've reworked this tutorial to take advantage of header capture.  This better demonstrates some of the advantages of using a separate gabc file and will make the tutorial more distinct from the snippet tutorial I'm working on (where this option isn't possible).