Tutorial updates
Addresses #50
We note that the tutorial is designed for 4.1 (i.e. the current TeXLive version), but that 4.2 users can make use of it by adding a bit of code to their TeX file.  What that code is and where to put it is noted in the tutorial at the point where it is first needed.