Documentation clean-up
"End of line" and "Custos" are split into separate sections in the documentation and the wording cleaned up for both.
In "End of line" the `<nlba>` tags are documented
In "Custos" the setting for automatic custos generation and how to override it at manual line breaks is documented.
Clarify that manual custos can be inserted anywhere, not just at end of score.